+++ src/main/java/org/infernus/idea/checkstyle/config/PluginConfiguration.java
@@ -1,17 +1,16 @@
 package org.infernus.idea.checkstyle.config;
 
+import com.intellij.openapi.components.ServiceManager;
+import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.project.Project;
 import net.jcip.annotations.Immutable;
 import org.infernus.idea.checkstyle.model.ConfigurationLocation;
+import org.infernus.idea.checkstyle.model.ConfigurationLocationFactory;
 import org.infernus.idea.checkstyle.model.ScanScope;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.Collections;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Objects;
-import java.util.SortedSet;
-import java.util.TreeSet;
+import java.util.*;
 import java.util.stream.Collectors;
 
 
@@ -27,7 +26,7 @@ public class PluginConfiguration {
     private final boolean copyLibs;
     private final SortedSet<ConfigurationLocation> locations;
     private final List<String> thirdPartyClasspath;
-    private final SortedSet<ConfigurationLocation> activeLocations;
+    private final SortedSet<String> activeLocationDescriptors;
     private final boolean scanBeforeCheckin;
     private final String lastActivePluginVersion;
 
@@ -37,7 +36,7 @@ public class PluginConfiguration {
                         final boolean copyLibs,
                         @NotNull final SortedSet<ConfigurationLocation> locations,
                         @NotNull final List<String> thirdPartyClasspath,
-                        @NotNull final SortedSet<ConfigurationLocation> activeLocations,
+                        @NotNull final SortedSet<String> activeLocationDescriptors,
                         final boolean scanBeforeCheckin,
                         @Nullable final String lastActivePluginVersion) {
         this.checkstyleVersion = checkstyleVersion;
@@ -46,24 +45,11 @@ public class PluginConfiguration {
         this.copyLibs = copyLibs;
         this.locations = Collections.unmodifiableSortedSet(locations);
         this.thirdPartyClasspath = Collections.unmodifiableList(thirdPartyClasspath);
-        this.activeLocations = activeLocations.stream()
-		        .map(e -> constrainActiveLocation(locations, e))
+        this.activeLocationDescriptors = activeLocationDescriptors.stream()
                 .filter(Objects::nonNull)
 		        .collect(Collectors.toCollection(TreeSet::new));
         this.scanBeforeCheckin = scanBeforeCheckin;
         this.lastActivePluginVersion = lastActivePluginVersion;
-    }
-
-    private ConfigurationLocation constrainActiveLocation(
-            @NotNull final SortedSet<ConfigurationLocation> sourceLocations,
-            @Nullable final ConfigurationLocation sourceActiveLocation) {
-        if (sourceActiveLocation != null && !sourceLocations.isEmpty()) {
-            return sourceLocations.stream()
-                    .filter(cl -> cl.equals(sourceActiveLocation))
-                    .findFirst()
-                    .orElse(null);
-        }
-        return null;
     }
 
     @NotNull
@@ -89,6 +75,27 @@ public class PluginConfiguration {
         return locations;
     }
 
+    private static final Logger LOG = Logger.getInstance(PluginConfiguration.class);
+
+    @NotNull
+    public Optional<ConfigurationLocation> findByDescriptor(@NotNull final String locationDescriptorToFind,
+                                                            @NotNull final Project project) {
+        // This is a horrid hack to get around IDEA resolving $PROJECT_DIR$ in descriptors, hence we need to do the whole load
+        // logic to make things consistent
+        // Ideally we should switch from descriptors to GUIDs or similar to avoid such things, although we'd
+        // still need the logic for legacy values
+        ConfigurationLocation resolvedConfiguration = configurationLocationFactory(project).create(project, locationDescriptorToFind);
+
+        return locations.stream()
+                .filter(location -> location.getDescriptor().equals(resolvedConfiguration.getDescriptor()))
+                .limit(1)
+                .findFirst();
+    }
+
+    private ConfigurationLocationFactory configurationLocationFactory(final Project project) {
+        return ServiceManager.getService(project, ConfigurationLocationFactory.class);
+    }
+
     @NotNull
     public List<String> getThirdPartyClasspath() {
         return thirdPartyClasspath;
@@ -99,8 +106,17 @@ public class PluginConfiguration {
         return lastActivePluginVersion;
     }
 
-    public SortedSet<ConfigurationLocation> getActiveLocations() {
-        return this.activeLocations;
+    public SortedSet<String> getActiveLocationDescriptors() {
+        return this.activeLocationDescriptors;
+    }
+
+    @NotNull
+    public SortedSet<ConfigurationLocation> getActiveLocations(@NotNull final Project project) {
+        return getActiveLocationDescriptors().stream()
+                .map(activeLocationDescriptor -> findByDescriptor(activeLocationDescriptor, project))
+                .filter(Optional::isPresent)
+                .map(Optional::get)
+                .collect(Collectors.toCollection(TreeSet::new));
     }
 
     public boolean isScanBeforeCheckin() {
@@ -139,7 +155,7 @@ public class PluginConfiguration {
                 && Objects.equals(copyLibs, otherDto.copyLibs)
                 && Objects.equals(locations, otherDto.locations)
                 && Objects.equals(thirdPartyClasspath, otherDto.thirdPartyClasspath)
-                && Objects.equals(activeLocations, otherDto.activeLocations)
+                && Objects.equals(activeLocationDescriptors, otherDto.activeLocationDescriptors)
                 && Objects.equals(scanBeforeCheckin, otherDto.scanBeforeCheckin)
                 && Objects.equals(lastActivePluginVersion, otherDto.lastActivePluginVersion);
     }
@@ -147,7 +163,7 @@ public class PluginConfiguration {
     @Override
     public int hashCode() {
         return Objects.hash(checkstyleVersion, scanScope, suppressErrors, copyLibs, locations, thirdPartyClasspath,
-                activeLocations, scanBeforeCheckin, lastActivePluginVersion);
+                activeLocationDescriptors, scanBeforeCheckin, lastActivePluginVersion);
     }
 
 }